Two camera calibration with not useful rectifing
